requirements.txt
alembic
asyncpg
fastapi[all]
fastapi-users[sqlalchemy]
python-dotenv
sqlalchemy
celery[redis]
